🐛 Festivities were marked as nobodies
In the translation from minizinc to tomatic representation
festivities were left as nobody/ningu which can be editable.
Backtracker marked such cells as festiu/festivity, which
the pbx scheduler knows is not a torn that should be open
and the editor knows (or should know) that it is not editable.