♻ mem->file: RELAY: third, use the new structure
Ongoing refactor to swap two equivalent implementations.

Third, RELAY on the new structure instead the old one.
We should replace all getters of the old structure.
Having the same information it should do no harm.

Any errors in the Duppe and Fill steps will be spotted here.

Indeed here i noticed that the 0 case, had to check
whether the file does not exist and return an empty
list if not.