✅ tests made independent of the generated id
For next test we would like to split the logs by operator.
If we do so, computing a sequential odoo_id won't be
so easy. It would imply persist the counter or read all
the operator logs, which is slow.

Foreseeing a non sequential id, let add_incomming_call
return the specified format which included the generated id
and use it to tweek the expectation in tests.