✅🐛 test__get_calls__ids_are_unique_among_operators
Because the id was the length of the log, and now
log is splitted by operator, calls for each operator
shared ids.

Now ids are generated by a random function of 128 bits.

This spotted an error in the previous test, since i
asserted odoo_id1 (alice's) for barbara's call.
Both were one and it didn't fail.