✨ RgbColor instead of pydantic Color
Pydantic's Color extra type is a class and gives problems
with yaml safe serialization.
Using a regexp constrained string instead.

Class serialization is blocked by safe yaml loader/dumper
for security reasons.

Even if we could, it would require adding serializers and
constructors and defining the type as implicit, which
less clear than defining the type as a constrained string.